signal-singapore-23/requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
discord-interactions==0.4.0
docopt==0.6.2
enum34==1.1.10
Flask==2.2.2
Flask-Cors==3.0.10
frozenlist==1.3.3
httpserver==1.1.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.4
ngrok==0.12.0
ngrok-api==0.10.0
pycparser==2.21
PyJWT==2.7.0
PyNaCl==1.5.0
python-dotenv==0.21.0
pytz==2023.3
requests==2.25.1
six==1.16.0
twilio==8.10.1
urllib3==1.26.13
Werkzeug==2.2.2
yarl==1.9.2
